Move functions to assertiodetection.ts
